third_party/WebKit/LayoutTests/inspector/console/console-call-getter-on-proto-expected.txt

CONSOLE MESSAGE: line 21: [object Object]
Tests that calling getter on prototype will call it on the object.

console-call-getter-on-proto.html:21 B {value: 239}
    foo: 239
    value: 239
    __proto__: A
        constructor: ƒ ()
        foo: 239
        __proto__: Object
